requirements.txt
asgiref==3.8.1
Django==5.0.7
django-bootstrap5==24.2
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
platformshconfig==2.4.0
sqlparse==0.5.0
